pulling find_by_name_in_nested_ui out into a helper module
